+++ popperJs/src/index.ts
@@ -1,3 +1,0 @@
-export * from './modules/popperjsPositioner.module';
-export * from './plugins/positioner/components';
-export * from './plugins/positioner';