+++ src/plugins/positioner/default/defaultPositioner.interface.ts
@@ -1,0 +1,15 @@
+import {Positioner, PositionerOptions} from '../positioner.interface';
+
+/**
+ * Default positioner options
+ */
+export interface DefaultPositionerOptions extends PositionerOptions
+{
+}
+
+/**
+ * Public API for 'DefaultPositionerComponent'
+ */
+export interface DefaultPositioner extends Positioner
+{
+}