+++ src/plugins/positioner/index.ts
@@ -1,4 +1,4 @@
 export * from './positioner.interface';
 
-export * from './basic/basicPositioner.interface';
+export * from './default/defaultPositioner.interface';
 export * from './no/noPositioner.interface';