+++ src/plugins/positioner/types.ts
@@ -1,7 +1,6 @@
 import {InjectionToken} from '@angular/core';
 
 import {PositionerOptions} from './positioner.interface';
-export {ScrollTargetSelector} from './basic/types';
 
 /**
  * Constant used for accessing positioner plugin in NgSelect